refactor: 피드백 반영, 패키지 분리에 따른 Test 코드 패키지 분리
