--- src/test/java/kitchenpos/application/ServiceTestBase.java
@@ -1,17 +1,0 @@
-package kitchenpos.application;
-
-import kitchenpos.DatabaseCleanup;
-import org.junit.jupiter.api.BeforeEach;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-public class ServiceTestBase {
-    @Autowired
-    private DatabaseCleanup databaseCleanup;
-
-    @BeforeEach
-    void setUp() {
-        databaseCleanup.execute();
-    }
-}
